+++ crates/ryo3-std/src/net/pydantic.rs
@@ -1,5 +1,5 @@
 use pyo3::prelude::*;
-use pyo3::types::{PyDict, PyTuple};
+use pyo3::types::PyDict;
 use ryo3_pydantic::{GetPydanticCoreSchemaCls, interns};
 
 use super::{PySocketAddr, PySocketAddrV4, PySocketAddrV6};
@@ -16,16 +16,16 @@ macro_rules! impl_get_pydantic_core_schema_cls_for_type {
                 let core_schema = ryo3_pydantic::core_schema(py)?;
                 let schema = core_schema.call_method(interns::str_schema(py), (), None)?;
                 let validation_fn = cls.getattr(interns::_pydantic_validate(py))?;
-                let args = PyTuple::new(py, vec![&validation_fn, &schema])?;
                 let string_serialization_schema =
                     core_schema.call_method(interns::to_string_ser_schema(py), (), None)?;
-                let serialization_kwargs = PyDict::new(py);
-                serialization_kwargs
+                let plain_validator_kwargs = PyDict::new(py);
+                plain_validator_kwargs.set_item(interns::json_schema_input_schema(py), &schema)?;
+                plain_validator_kwargs
                     .set_item(interns::serialization(py), &string_serialization_schema)?;
                 core_schema.call_method(
-                    interns::no_info_wrap_validator_function(py),
-                    args,
-                    Some(&serialization_kwargs),
+                    interns::no_info_plain_validator_function(py),
+                    (&validation_fn,),
+                    Some(&plain_validator_kwargs),
                 )
             }
         }